Move some files that are specific to the react-reconciler from shared
Individual renderers are allowed to deep require into the reconciler.
