+++ packages/legacy-events/__tests__/ResponderEventPlugin-test.internal.js
@@ -1378,7 +1378,7 @@ describe('ResponderEventPlugin', () => {
     // ResponderEventPlugin uses `getLowestCommonAncestor`
     const React = require('react');
     const ReactTestUtils = require('react-dom/test-utils');
-    const ReactTreeTraversal = require('shared/ReactTreeTraversal');
+    const ReactTreeTraversal = require('react-reconciler/src/ReactTreeTraversal');
     const ReactDOMComponentTree = require('../../react-dom/src/client/ReactDOMComponentTree');
 
     class ChildComponent extends React.Component {
