+++ packages/react-reconciler/src/ReactFiberUnwindWork.js
@@ -21,7 +21,7 @@ import {
   SuspenseComponent,
   SuspenseListComponent,
 } from 'shared/ReactWorkTags';
-import {DidCapture, NoEffect, ShouldCapture} from 'shared/ReactSideEffectTags';
+import {DidCapture, NoEffect, ShouldCapture} from './ReactSideEffectTags';
 import {enableSuspenseServerRenderer} from 'shared/ReactFeatureFlags';
 
 import {popHostContainer, popHostContext} from './ReactFiberHostContext';
