+++ packages/react-reconciler/src/ReactUpdateQueue.js
@@ -94,7 +94,7 @@ import {
   enterDisallowedContextReadInDEV,
   exitDisallowedContextReadInDEV,
 } from './ReactFiberNewContext';
-import {Callback, ShouldCapture, DidCapture} from 'shared/ReactSideEffectTags';
+import {Callback, ShouldCapture, DidCapture} from './ReactSideEffectTags';
 
 import {debugRenderPhaseSideEffectsForStrictMode} from 'shared/ReactFeatureFlags';
